Improve ch-flexible-layout's component structure
