+++ src/components/flexible-layout/utils.ts
@@ -1,191 +1,197 @@
-import { inBetween } from "../../common/utils";
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { tabTypeToPart } from "../list/utils";
 import {
-  DraggableViewExtendedInfo,
-  DroppableArea,
-  WidgetDropInfo
-} from "./types";
+  FlexibleLayout,
+  FlexibleLayoutGroup,
+  FlexibleLayoutItem,
+  FlexibleLayoutItemExtended,
+  FlexibleLayoutLeaf,
+  FlexibleLayoutLeafInfo,
+  FlexibleLayoutLeafType,
+  FlexibleLayoutWidgetExtended
+} from "./internal/flexible-layout/types";
+import { ROOT_VIEW } from "../../common/utils";
 
-/**
- * [block-start, block-end, inline-start, inline-end]
- */
-type DroppableAreaSizes = [number, number, number, number];
+// Aliases
+type ItemExtended = FlexibleLayoutItemExtended<
+  FlexibleLayoutItem,
+  FlexibleLayoutLeafType
+>;
 
-// Custom vars
-const BLOCK_START = "--ch-flexible-layout-drop-area-block-start";
-const BLOCK_END = "--ch-flexible-layout-drop-area-block-end";
-const INLINE_START = "--ch-flexible-layout-drop-area-inline-start";
-const INLINE_END = "--ch-flexible-layout-drop-area-inline-end";
+type LeafExtended = FlexibleLayoutItemExtended<
+  FlexibleLayoutLeaf,
+  FlexibleLayoutLeafType
+>;
 
-/**
- * If the mouse position is in the interval [0%, 30%] the droppable area should
- * be displayed at the START position with half the size.
- */
-const START_HALF_THRESHOLD = 30; // In percentage
+type GroupExtended = FlexibleLayoutItemExtended<
+  FlexibleLayoutGroup,
+  FlexibleLayoutLeafType
+>;
 
-/**
- * If the mouse position is in the interval [70%, 100%] the droppable area
- * should be displayed at the END position with half the size.
- */
-const END_HALF_THRESHOLD = 100 - START_HALF_THRESHOLD; // In percentage
+export const createAndSetLeafInfo = (
+  flexibleLayoutLeaf: FlexibleLayoutLeaf,
+  renderedWidgets: Set<string>,
+  widgetsInfo: Map<string, FlexibleLayoutWidgetExtended>
+): FlexibleLayoutLeafInfo<FlexibleLayoutLeafType> => {
+  const leafId = flexibleLayoutLeaf.id;
+  const leafType = flexibleLayoutLeaf.type;
 
-const EDGE_SIZE = 0.5;
+  if (leafType === "single-content") {
+    const widget = flexibleLayoutLeaf.widget;
 
-const setProperty = (element: HTMLElement, property: string, value: number) =>
-  element.style.setProperty(property, `${value}px`);
+    // Mark the widget as rendered
+    renderedWidgets.add(widget.id);
 
-const inlineStart = (mainViewRect: DOMRect, rtl: boolean) =>
-  rtl ? mainViewRect.left + mainViewRect.width * EDGE_SIZE : mainViewRect.left;
+    // Store the widget info
+    widgetsInfo.set(leafId, { parentLeafId: leafId, info: widget });
 
-const inlineEnd = (
-  documentRect: DOMRect,
-  mainViewRect: DOMRect,
-  rtl: boolean
-) =>
-  rtl
-    ? documentRect.width - mainViewRect.right
-    : documentRect.width - (mainViewRect.left + mainViewRect.width * EDGE_SIZE);
+    return {
+      id: leafId,
+      type: leafType,
+      exportParts: "",
+      widget: widget
+    };
+  }
 
-const droppableAreaMap: {
-  [key in DroppableArea]: (
-    documentRect: DOMRect,
-    mainViewRect: DOMRect,
-    rtl: boolean
-  ) => DroppableAreaSizes;
-} = {
-  "block-start": (documentRect, mainViewRect) => [
-    mainViewRect.top,
-    documentRect.height - (mainViewRect.top + mainViewRect.height * EDGE_SIZE),
-    mainViewRect.left,
-    documentRect.width - mainViewRect.right
-  ],
+  let selectedWidgetId = flexibleLayoutLeaf.selectedWidgetId;
+  const widgets = flexibleLayoutLeaf.widgets;
+  const tabOrientation = flexibleLayoutLeaf.tabDirection;
+  const tabPosition = flexibleLayoutLeaf.tabPosition;
 
-  "block-end": (documentRect, mainViewRect) => [
-    mainViewRect.top + mainViewRect.height * EDGE_SIZE,
-    documentRect.height - (mainViewRect.top + mainViewRect.height),
-    mainViewRect.left,
-    documentRect.width - mainViewRect.right
-  ],
+  const exportParts =
+    tabTypeToPart[
+      `${tabOrientation}-${
+        tabPosition ?? "start"
+      }` as keyof typeof tabTypeToPart
+    ](widgets);
 
-  "inline-start": (documentRect, mainViewRect, rtl) => [
-    mainViewRect.top,
-    documentRect.height - (mainViewRect.top + mainViewRect.height),
-    inlineStart(mainViewRect, rtl),
-    inlineEnd(documentRect, mainViewRect, rtl)
-  ],
+  widgets.forEach(widget => {
+    if (widget.wasRendered || selectedWidgetId === widget.id) {
+      // Ensure proper initialization
+      widget.wasRendered = true;
 
-  "inline-end": (documentRect, mainViewRect, rtl) => [
-    mainViewRect.top,
-    documentRect.height - (mainViewRect.top + mainViewRect.height),
-    inlineStart(mainViewRect, !rtl),
-    inlineEnd(documentRect, mainViewRect, !rtl)
-  ],
+      renderedWidgets.add(widget.id);
+    }
 
-  center: (documentRect, mainViewRect) => [
-    mainViewRect.top,
-    documentRect.height - (mainViewRect.top + mainViewRect.height),
-    mainViewRect.left,
-    documentRect.width - mainViewRect.right
-  ]
+    // Store the widget info
+    widgetsInfo.set(widget.id, { parentLeafId: leafId, info: widget });
+  });
+
+  // If there is no widget selected by default, select one
+  if (selectedWidgetId == null) {
+    const selectedWidget =
+      widgets[tabOrientation === "block" ? widgets.length - 1 : 0];
+    selectedWidgetId = selectedWidget.id;
+    selectedWidget.wasRendered = true;
+
+    // Mark the widget as rendered
+    renderedWidgets.add(selectedWidgetId);
+  }
+
+  return {
+    id: leafId,
+    exportParts,
+    closeButtonHidden: flexibleLayoutLeaf.closeButtonHidden ?? false,
+    dragOutsideDisabled: flexibleLayoutLeaf.dragOutsideDisabled ?? false,
+    selectedWidgetId: selectedWidgetId,
+    showCaptions: flexibleLayoutLeaf.showCaptions ?? true,
+    sortable: flexibleLayoutLeaf.sortable ?? true,
+    tabDirection: tabOrientation,
+    tabPosition: flexibleLayoutLeaf.tabPosition,
+    type: leafType,
+    widgets: widgets
+  };
 };
 
-let lastDroppableArea: DroppableArea;
-let lastViewId: string;
+const addCustomBehavior = (
+  item: FlexibleLayoutItem,
+  layoutSplitterParts: Set<string>
+) => {
+  if (item.dragBar?.part) {
+    layoutSplitterParts.add(item.dragBar?.part);
+  }
+};
 
-export const handleWidgetDrag =
-  (
-    draggableView: DraggableViewExtendedInfo,
-    droppableAreaRef: HTMLElement,
-    rtl: boolean
-  ) =>
-  (event: MouseEvent) => {
-    event.stopPropagation(); // Prevents the remove of the droppable area
-
-    // - - - - - - - - - - - DOM read operations - - - - - - - - - - -
-    const documentRect = document.documentElement.getBoundingClientRect();
-    const mainViewRect = draggableView.mainView.getBoundingClientRect();
-    const positionX = event.clientX; // Mouse position X
-    const positionY = event.clientY; // Mouse position Y
-
-    const distanceToTheLeftEdge = rtl
-      ? mainViewRect.width - (positionX - mainViewRect.left)
-      : positionX - mainViewRect.left;
-    const distanceToTheTopEdge = positionY - mainViewRect.top;
-
-    const relativePositionX =
-      (distanceToTheLeftEdge / mainViewRect.width) * 100;
-    const relativePositionY =
-      (distanceToTheTopEdge / mainViewRect.height) * 100;
-
-    let droppableArea: DroppableArea;
-
-    // Block start (Most likely droppable area)
-    if (
-      relativePositionY <= START_HALF_THRESHOLD &&
-      inBetween(relativePositionY, relativePositionX, 100 - relativePositionY)
-    ) {
-      droppableArea = "block-start";
-    }
-
-    // Inline End (second most likely droppable area)
-    else if (
-      relativePositionX >= END_HALF_THRESHOLD &&
-      inBetween(100 - relativePositionX, relativePositionY, relativePositionX)
-    ) {
-      droppableArea = "inline-end";
-    }
-
-    // Inline Start
-    else if (
-      relativePositionX <= START_HALF_THRESHOLD &&
-      inBetween(relativePositionX, relativePositionY, 100 - relativePositionX)
-    ) {
-      droppableArea = "inline-start";
-    }
-
-    // Block end
-    else if (
-      relativePositionY >= END_HALF_THRESHOLD &&
-      inBetween(100 - relativePositionY, relativePositionX, relativePositionY)
-    ) {
-      droppableArea = "block-end";
-    }
-
-    // Center
-    else {
-      droppableArea = "center";
-    }
-
-    // If the droppable area did not change, there is no need to update the DOM
-    if (
-      lastViewId === draggableView.viewId &&
-      lastDroppableArea === droppableArea
-    ) {
-      return;
-    }
-    lastDroppableArea = droppableArea;
-    lastViewId = draggableView.viewId;
-
-    const droppableAreaSizes = droppableAreaMap[droppableArea](
-      documentRect,
-      mainViewRect,
-      rtl
-    );
-
-    // - - - - - - - - - - - DOM write operations - - - - - - - - - - -
-    setProperty(droppableAreaRef, BLOCK_START, droppableAreaSizes[0]);
-    setProperty(droppableAreaRef, BLOCK_END, droppableAreaSizes[1]);
-
-    setProperty(droppableAreaRef, INLINE_START, droppableAreaSizes[2]);
-    setProperty(droppableAreaRef, INLINE_END, droppableAreaSizes[3]);
+export const addNewLeafToInfo = (
+  leaf: FlexibleLayoutLeaf,
+  parentItem: FlexibleLayoutGroup,
+  itemsInfo: Map<string, ItemExtended>,
+  renderedWidgets: Set<string>,
+  widgetsInfo: Map<string, FlexibleLayoutWidgetExtended>
+) => {
+  const flexibleLeafExtended: LeafExtended = {
+    item: leaf,
+    parentItem: parentItem,
+    leafInfo: createAndSetLeafInfo(leaf, renderedWidgets, widgetsInfo)
   };
 
-export const removeDroppableAreaStyles = (droppableAreaRef: HTMLElement) => {
-  lastDroppableArea = undefined;
-  lastViewId = undefined;
-  droppableAreaRef.removeAttribute("style");
+  itemsInfo.set(leaf.id, flexibleLeafExtended);
 };
 
-export const getWidgetDropInfo = (): WidgetDropInfo | undefined =>
-  lastDroppableArea === undefined
-    ? undefined
-    : { dropAreaTarget: lastDroppableArea, viewIdTarget: lastViewId };
+const updateFlexibleSubModels = (
+  flexibleLayoutItems: FlexibleLayoutItem[],
+  itemsInfo: Map<string, ItemExtended>,
+  layoutSplitterParts: Set<string>,
+  renderedWidgets: Set<string>,
+  widgetsInfo: Map<string, FlexibleLayoutWidgetExtended>,
+  parentItem: FlexibleLayoutGroup
+) => {
+  flexibleLayoutItems.forEach(flexibleItem => {
+    // Group
+    if ((flexibleItem as FlexibleLayoutGroup).items != null) {
+      const group = flexibleItem as FlexibleLayoutGroup;
+
+      const flexibleItemExtended: GroupExtended = {
+        item: group,
+        parentItem: parentItem
+      };
+      itemsInfo.set(group.id, flexibleItemExtended);
+
+      layoutSplitterParts.add(group.id);
+
+      updateFlexibleSubModels(
+        group.items, // Subitems
+        itemsInfo,
+        layoutSplitterParts,
+        renderedWidgets,
+        widgetsInfo,
+        group
+      );
+    }
+    // Leaf
+    else {
+      addNewLeafToInfo(
+        flexibleItem as FlexibleLayoutLeaf,
+        parentItem,
+        itemsInfo,
+        renderedWidgets,
+        widgetsInfo
+      );
+    }
+
+    // Custom behaviors
+    addCustomBehavior(flexibleItem, layoutSplitterParts);
+  });
+};
+
+export const updateFlexibleModels = (
+  flexibleLayout: FlexibleLayout,
+  itemsInfo: Map<string, ItemExtended>,
+  layoutSplitterParts: Set<string>,
+  renderedWidgets: Set<string>,
+  widgetsInfo: Map<string, FlexibleLayoutWidgetExtended>
+) =>
+  updateFlexibleSubModels(
+    flexibleLayout.items,
+    itemsInfo,
+    layoutSplitterParts,
+    renderedWidgets,
+    widgetsInfo,
+    ROOT_VIEW // Root item
+  );
+
+export const getLeafInfo = (
+  itemsInfo: Map<string, ItemExtended>,
+  leafId: string
+): FlexibleLayoutLeafInfo<FlexibleLayoutLeafType> =>
+  (itemsInfo.get(leafId) as LeafExtended).leafInfo;
